Algebra

The construction of an algebra is the first step for executing code in geometric algebra with MinGal, as it defines the main characteristics for creating the multivector space and its operations.

The function Algebra is also described below and is the main function used to initialize spaces.

Main.MinGal.Algebra — Type

Algebra(p, [q], [r], [symbols], [type])::Algebra

Main function for creating your Algebra and adding its basis blades to REPL. Constructor Function of an algebraic object with parameters p, q, r, R^{p, q, r}, and its multivector space. If not defined, the last two parameters are automatically calculated as canonical.

Arguments

p::Int : Represents the ammount of positive dimensions

q::Int : Represents the ammount of negative dimensions

r::Int : Represents the ammount of null dimensions

symbols::Vector{String} : Array of primary symbols for the Algebra

type::String : String for manually selecting "full", "min" or "special" Algebras

Return

Returns the created Algebra object.

The function above is responsible for creating the algebra used throughout the code or within the function where it was initialized, as well as instantiating the basis blades and their operations. The canonical blades can be accessed using the symbol $e$ followed by the number representing them. Additionally, it is also possible to store the algebra in a variable and use it to access the blades.

julia> al = Algebra(2, 1)
julia> e1 + e1e2
julia> al.e3 * al.e2e3

For this example, the created symbols are:

Grade-0 blade (Scalar): id the scalar as a multivector type

Grade-1 blades (Vectors): e1, e2, and e3

Grade-2 blades (Bivectors): e1e2, e1e3, and e2e3

Grade-3 blade (Trivector): e1e2e3 or eI the pseudosclalar

In degenerate algebras, that is, with a component $r$ different from 0, the first symbol, if not otherwise specified, starts with e0 instead of e1, and the counting of null dimensions begins from that.

julia> al = Algebra(2, 0, 1)

julia> e0 * e0
0.0

Types

An algebra can be 'complete', as in AlgebraFull, or 'minimal', as in AlgebraMin. The difference lies in the amount of information stored and preprocessed during its construction, which is chosen automatically or via the type = "full" and type = "min" parameters in the definition. Additionally, for very large algebras, there is another option, type = "special", in which no symbols for the basis blades are constructed, as doing so would be highly inefficient.

You can choose this by setting the parameter, or leave it in automatic mode for MinGal to handle. This allows generating algebras in very high dimensions, although with more limited capabilities.

In AlgebraFull, all symbols for the generated basis blades are created. In AlgebraMin, only the blades with grade equal to 1, or no blades at all, are generated, depending on the size.

julia> Algebra(1000)
julia> e1 + e2 + e1000

Describe

To obtain more information about the generated algebra, such as its metric, symbols, blades, basis, and more, you can use the function describe, as shown in the following example:

julia> describe(al)Algebra:
- p: 2
- q: 0
- r: 1
- basis_bit_order: ["1", "e0", "e1", "e0e1", "e2", "e0e2", "e1e2", "e0e1e2"]
- metric: Int8[0, 1, 1]

Custom Symbols

An interesting possibility for creating algebras is symbol customization, where you can replace the canonical basis blades with the symbols of your choice, and MinGal will automatically concatenate them to form higher grade blades.

julia> Complex = Algebra(0, 1, 0, ["i"])

julia> i*i
- 1

julia> (1 + 2*i)^2 * (2 - i)
- 2 + 11.0*i

julia> Algebra(2, 0, 1, ["a", "b", "c"])

julia> a*a
0.0

julia> (a + 2abc) \ bc
- 2.0*a

julia> eI
1.0*abc

Multiple Algebras

In order to create and use more than one algebra to perform operations, there are some important points and precautions to take into consideration, as well as best practices and the limitations of MinGal.

Each time the Algebra function is executed, an Algebra object and its symbols are created in the REPL so they can be used both with dot notation and by directly writing the symbols. However, conflicts occur when two identical symbols are created, as shown in the example below:

julia> R2 = Algebra(2,0,0,["i", "j"])
julia> Complex = Algebra(0,1,0,["i"])

In this case, it is always recommended to use dot notation for any operations as a programming best practice. However, it is still possible to use, for example, the symbols created as i and j. The value assigned to j is the one created by R2, since there is no conflict. However, the symbol i takes on the value of Complex because it overwrote the previous definition.

Thus, to resolve this conflict and give you full control over which algebra you are working with, you can use dot notation or use the change_algebra command to safely switch the algebra’s symbols. The same applies to the shared symbols id and eI.

Warning - change_algebra

In high-performance situations, this function is not recommended because it recreates the same symbols again, resulting in suboptimal time complexity; prefer using dot notation. If this is not an issue, then it can be used normally.

julia> R2 = Algebra(2,0,0,["i", "j"])
julia> Complex = Algebra(0,1,0,["i"])

julia> i*i
- 1

julia> j*j
1

julia> eI
1.0*i

julia> Complex.i * Complex.i
- 1

julia> R2.i * R2.i
1

julia> change_algebra(R2)

julia> i*i
1

julia> eI
1.0*ij