References

Here you will find all the docstring documentation for the implemented functions.

Main.MinGal.Algebra — Type

Algebra(p, [q], [r], [symbols], [type])::Algebra

Main function for creating your Algebra and adding its basis blades to REPL. Constructor Function of an algebraic object with parameters p, q, r, R^{p, q, r}, and its multivector space. If not defined, the last two parameters are automatically calculated as canonical.

Arguments

p::Int : Represents the ammount of positive dimensions

q::Int : Represents the ammount of negative dimensions

r::Int : Represents the ammount of null dimensions

symbols::Vector{String} : Array of primary symbols for the Algebra

type::String : String for manually selecting "full", "min" or "special" Algebras

Return

Returns the created Algebra object.

Main.MinGal.AlgebraFull — Type

AlgebraFull(p, q, r, symbols, basis, basis_bit_order, metric, max)

A structure to define an algebra to be worked with its respective dimensions and canonical vectors.

Fields

p::Int : Represents the ammount of positive dimensions

q::Int : Represents the ammount of negative dimensions

r::Int : Represents the ammount of null dimensions

symbols::Vector{String} : Array of primary symbols for the Algebra

basis::Vector{String} : Array of all symbols for the Algebra, normal order

basis_bit_order::Vector{String} : Array of symbols for the Algebra, bit order

blades::Dict{Symbol, Any} : Array of all multivectors generated for this algebra

metric::Vector{Int8} : Another way of representing the algebra signature

max::Integer : max number of Algebra elements, the same as 2^(p+q+r)

Main.MinGal.AlgebraMin — Type

AlgebraMin(p, q, r, symbols, metric, max)

A structure to define an algebra to be worked with its respective dimensions and canonical vectors.

Fields

p::Int : Represents the ammount of positive dimensions

q::Int : Represents the ammount of negative dimensions

r::Int : Represents the ammount of null dimensions

symbols::Vector{String} : Array of primary symbols for the Algebra

blades::Dict{Symbol, Any} : Array of all multivectors generated for this algebra

metric::Vector{Int8} : Another way of representing the algebra signature

max::Int : max number of Algebra elements, the same as 2^(p+q+r)

Main.MinGal.Blade — Type

Blade(bitmap, scalar, al::Algebra)::Blade ||
Blade(mv::Multivector)::Blade ||
Blade(k::Number, al::Algebra)::Blade

Creates a Blade based on bitmap and Scalars or Converts a multivector into a blade. It may throw an error

Arguments

bitmap::Integer : The bitmap that specifies what basis vectors are present in this blade

scalar::Number : The scalar of the basis blade.

al::Algebra : The current algebra on use.

||

mv::Multivector

||

k::Number

al::Algebra : The current algebra on use.

Return

Returns a Blade.

Main.MinGal.Blade — Type

Blade(blade_array)

Struct that creates the Blade object. The bitmap that specifies what basis vectors are present in this blade and The scalar of the basis blade.

Arguments

basis blades and their scalars.

Main.MinGal.Multivector — Type

Multivector(base_vectors, scalars, al::Algebra) ||
Multivector(bl::Blade) ||
Multivector(blades::Array{Blade}, al::Algebra)

Constructor function for creating multivectors.

Arguments

base_vectors::Vector{Integer} : An array of integers, representing the actual basis blade that exists in this object in bit order.

scalars::Vector{Number} : An array of Floats, representing the scalars of each basis blade in bit order.

al::Algebra : The current algebra on use.

||

bl::Blade : A Blade, to convert to Multivector.

||

blades::GAVector : An array of Blades, to convert to Multivector.

al::Algebra : The current algebra on use.

Return

Returns a Multivector.

Main.MinGal.Multivector — Type

Multivector(blade_array)

Struct that creates the multivector object.

Arguments

blade_array::SparseArrays.SparseVector{Number, Integer} : An sparse vector with the internal values of

basis blades and their scalars.

Main.MinGal.binary_index_map — Method

binary_index_map(base, basis)::Dict{Int, Int}

Function that maps the index of normal and bit order.

Arguments

base::Vector{String} : An array of strings to be combined.

basis::Vector{String} : An array of the final combination of strings.

Return

Returns a dict with all the mapped values.

Main.MinGal.bitmap — Method

bitmap(bl::Blade)::Integer

Returns the bitmap of the Blade

Arguments

bl::Blade

Return

The bitmap

Main.MinGal.canon_basis — Method

canon_basis(symbols)::Vector{String}

Function that lists all the combinations of canonical vectors in a given Algebra.

Arguments

symbols::Vector{String} : An array of strings to be combined.

Return

Returns a list with all combinations of the elements, forming the basis of the multivector space.

Main.MinGal.canon_basis_bit_order — Method

canon_basis_bit_order(symbols)::Vector{String}

Function that lists all the combinations of canonical vectors in a given Algebra in bit order.

Arguments

symbols::Vector{String} : An array of strings to be combined.

Return

Returns a list with all combinations of the elements, forming the basis of the multivector space, in bit order.

Main.MinGal.canon_symbols — Function

canon_symbols(p, [q], [r])::Vector{String}

Function that writes the canonical vector space symbols, given the parameters p, q and r for definition

Arguments

p::Int : Represents the ammount of positive dimensions

q::Int : Represents the ammount of negative dimensions

r::Int : Represents the ammount of null dimensions

Return

Return an array of strings with all the necessary elements for this space.

Main.MinGal.canonical_basis — Function

canonical_basis(al::Algebra)::GAVector

Function that returns an iterable list with only the canonical basis blades

Arguments

al::Algebra

Return

A Vector (GAVector) with only the canonical basis blades

Main.MinGal.canonical_reordering_sign — Method

canonical_reordering_sign(ei, ej)::Integer

Function that computes the reordering of a Blade to get into canonical order.

Arguments

ei::Blade

ej::Blade

Return

The sign value, either 1 or -1.

Main.MinGal.chain — Function

chain(coeff::Vector, al::Algebra)::GAType

Function that 'chains' the canonical basis with a vector of coefficients. It sums the coeff[i]*canonical[i] with the rest, good for translating matrices lines into a GA object.

Arguments

coeff::Vector

al::Algebra

Return

a single GAType, the chained value

Main.MinGal.change_algebra — Method

change_algebra(al::Algebra)

Change the algebra REPL symbols on use. Not recomended because of time complexity. Prefer using dot notation when working with multiple algebras.

Arguments

al::Algebra

Main.MinGal.clifford_conjugation — Method

clifford_conjugation(ei::GAType)::GAType

Function that computes the Clifford Conjugation of a multivector and return its result.

Arguments

ei::Multivector

Return

The Clifford Conjugate of ei.

Main.MinGal.clifford_norm — Method

clifford_norm(ei::GAType)::Number

Function that computes the clifford norm of a GAType.

Arguments

ei::GAType

Return

The clifford norm of ei.

Main.MinGal.conjugate — Method

conjugate(ei::GAType)::GAType

Function that computes the conjugate of a multivector and return its result.

Arguments

ei::Multivector

Return

The conjugate of ei.

Main.MinGal.conjugate_norm — Method

conjugate_norm(ei::GAType)::Number

Function that computes the conjugate norm of a GAType. Also another possible norm for degenerate algebras.

Arguments

ei::GAType

Return

The conjugate norm of ei.

Main.MinGal.create_algebra — Function

create_algebra(p, [q], [r], [symbols])

Constructor Function of an algebraic object with signature p, q, r. If not defined, the symbols for the algebra are automatically calculated as canonical.

Arguments

p::Int : Represents the ammount of positive dimensions

q::Int : Represents the ammount of negative dimensions

r::Int : Represents the ammount of null dimensions

symbols::Vector{String} : Array of primary symbols for the Algebra

Return

Returns the created Algebra object.

Main.MinGal.create_algebra_min — Function

create_algebra_min(p, [q], [r])

Constructor Function of an algebraic object with signature p, q, r. There are no symbos in the "min" algebra representation.

Arguments

p::Int : Represents the ammount of positive dimensions

q::Int : Represents the ammount of negative dimensions

r::Int : Represents the ammount of null dimensions

Return

Returns the created Algebra object.

Main.MinGal.create_special_symbols — Method

create_special_symbols()

Create and add to REPL the special symbols for this Algebra, such as the "id" for scalar GAType grade 0 and "eI" for the pseudoscalar of the space.

Main.MinGal.create_symbols — Method

create_symbols(stringSymbols)

Create and add to REPL all the (custom or basis) symbols for this Algebra.

Arguments

string_symbols::Vector{String} : An array with all the custom or basis symbols.

Main.MinGal.create_symbols_min — Method

create_symbols_min(stringSymbols)

Create and add to REPL the (custom or basis) symbols for this Algebra. But only the canonical basis blades symbols no combination between then.

Arguments

string_symbols::Vector{String} : An array with all the custom or basis symbols.

Main.MinGal.describe — Method

describe(al::Algebra)

Describe function for showing the Algebra function.

Arguments

al::Algebra : The algebra for printing

Main.MinGal.dual — Method

dual(ei::GAType)::GAType

Function that computes the dual of a multivector and return its result. Just be careful with parameter r > 0. It uses Polarity algorithm for invertible algebras and hodge dual for degenerate algebras.

Arguments

ei::Multivector

Return

The dual of ei.

Main.MinGal.euclidean_norm — Method

euclidean_norm(ei::GAType)::Number

Function that computes the euclidean coefficients norm of a GAType. It is also good for degenerate algebras because the metric is mostly ignored.

Arguments

ei::GAType

Return

The euclidean norm of ei.

Main.MinGal.euclidean_scalar_product — Method

euclidean_scalar_product(ei, ej)::Number

Function that returns the Euclidean Scalar Product between two GAType.

Arguments

ei::GAType

ej::GAType

Return

The result product

Main.MinGal.exp_ga — Method

exp_ga(ei::GAType)::GAType

Function that computes the exponential of elements that squares to a scalar. A more generic function was not yet implemented.

Arguments

ei::GAType

Return

The result GAType.

Main.MinGal.geometric_product — Method

geometric_product(ei, ej)::Multivector

Function that returns the Geometric Product between two blades.

Arguments

ei::Blade

ej::Blade

Return

The result Blade.

Main.MinGal.geometric_product — Method

geometric_product(ei::GAType, ej::GAType)::GAType

Function that computes the geometric product of two multivectors and return its result.

Arguments

ei::GAType

ej::GAType

Return

The result GAType.

Main.MinGal.get_scalar — Method

get_scalar(mv, k)::Number ||
get_scalar(mv, ei)::Number

Function that returns the scalar value in index k from GAType mv. The index k follows the bit order basis.The ei value represents the blade that you want the scalar from.

Arguments

mv::GAType : A Blade or a Multivector.

k::Integer : Index in bit order basis

||

mv::GAType : A Blade or a Multivector.

ei::Blade : Blade for Index

Return

The scalar value.

Main.MinGal.grade — Method

grade(bl::Blade)::Integer ||
grade(mv::Multivector)::Integer

Function that returns the grade of the Blade or multivector.

Arguments

bl::Blade

||

mv::Multivector

Return

An integer, the grade of the blade or multivector.

Main.MinGal.grade_involution — Method

grade_involution(ei::GAType)::GAType

Function that computes the grade involuction of a multivector and return its result.

Arguments

ei::Multivector

Return

The grade involution of ei.

Main.MinGal.grade_minus — Method

grade_minus(bl::Blade)::Integer

Function that returns the numbers of basis vectors in a basis blade that square to -1.

Arguments

bl::Blade

Return

An integer, the grade- of the blade or multivector.

Main.MinGal.grade_null — Method

grade_null(bl::Blade)::Integer

Function that returns the numbers of basis vectors in a basis blade that square to 0.

Arguments

bl::Blade

Return

An integer, the grade0 of the blade or multivector.

Main.MinGal.grade_plus — Method

grade_plus(bl::Blade)::Integer

Function that returns the numbers of basis vectors in a basis blade that square to +1.

Arguments

bl::Blade

Return

An integer, the grade+ of the blade or multivector.

Main.MinGal.grade_projection — Method

grade_projection(bl, k)::Blade ||
grade_projection(mv, k)::Blade

Function that returns the grade Projection between a Blade and an Integer. In the multivector form, it requires that it has only one blade.

Arguments

bl::Blade

k::Integer : The grade for projection

||

mv::Blade

k::Integer : The grade for projection

Return

The result Blade. It might be the 1D blade "1"

Main.MinGal.grade_selection — Method

grade_selection(ei::GAType, k::Number)::GAType

Function that retrieves the k grade part of a GAType. Duplicated, since grade_projection does the same, but a new algorithm and more generalized.

Arguments

ei::Multivector

k::Number : The selected grade.

Return

The result GAType.

Main.MinGal.has_key — Method

has_key(mv, k)::Bool ||
has_key(mv, ei)::Bool

Function that checks if the GAType has a blade.

Arguments

mv::GAType

k::Integer : Index

||

mv::GAType

ei::Blade : Index

Return

True, if yes, False, if not.

Main.MinGal.inner_product — Method

inner_product(ei, ej)::Multivector

Function that returns the Inner Product between two blades.

Arguments

ei::Blade

ej::Blade

Return

The result Blade.

Main.MinGal.inner_product — Method

inner_product(ei::GAType, ej::GAType)::GAType

Function that computes the outer product of two multivectors and return its result.

Arguments

ei::GAType

ej::GAType

Return

The result GAType.

Main.MinGal.invert — Method

invert(ei::GAType)::GAType

Function that computes the inverse of a multivector (mostly versors and blades supported!) and return its result. May throw an error if not invertible.

Arguments

ei::GAType

Return

The inverse of ei.

Main.MinGal.is_blade — Method

is_blade(ei::GAType)::Bool

Function that checks where a GAType is a Blade or not.

Arguments

ei::GAType

Return

True if it's a Blade, False otherwise.

Main.MinGal.is_scalar — Method

is_scalar(ei::GAType)::Bool

Function that checks where a GAType is a Scalar or not.

Arguments

ei::GAType

Return

True if it's a Scalar, False otherwise.

Main.MinGal.left_contraction — Method

left_contraction(ei::GAType, ej::GAType)::GAType

Function that computes the left contraction of two multivectors and return its result.

Arguments

ei::GAType

ej::GAType

Return

The result GAType.

Main.MinGal.multivector_sub — Method

multivector_sub(mi::GAType, mj::GAType)::GAType

Function that subtracts two multivectors and return its result.

Arguments

mi::GAType

mj::GAType

Return

The result GAType.

Main.MinGal.multivector_sum — Method

multivector_sum(mi::GAType, mj::GAType)::GAType

Function that sums two GAType and return its result.

Arguments

mi::GAType

mj::GAType

Return

The result Multivector.

Main.MinGal.norm — Method

norm(ei::GAType)::Number

Function that computes the reverse norm in absolute value (conjugate norm) for a GAType on usual algebras. For degenerate algebras, computes euclidean norm.

Arguments

ei::GAType

Return

The norm of ei.

Main.MinGal.outer_product — Method

outer_product(ei, ej)::Multivector

Function that returns the Outer Product between two blades.

Arguments

ei::Blade

ej::Blade

Return

The result Blade.

Main.MinGal.outer_product — Method

outer_product(ei::GAType, ej::GAType)::GAType

Function that computes the outer product of two multivectors and return its result.

Arguments

ei::GAType

ej::GAType

Return

The result GAType.

Main.MinGal.product_by_scalar — Method

product_by_scalar(mv::GAType, k::Number)::GAType

Function that calculates the product by scalar between a blade and a scalar.

Arguments

mv::GAType

k::Number

Return

The result GAType.

Main.MinGal.regressive_product — Method

regressive_product(ei::GAType, ej::GAType)::GAType

Function that computes the regressive product of two multivectors and return its result.

Arguments

ei::GAType

ej::GAType

Return

The result GAType.

Main.MinGal.reverse_norm — Method

reverse_norm(ei::GAType)::Number

Function that computes the reverse norm of a GAType on usual algebras. In practice the reverse norm is useful, especially due to its possible negative sign

Arguments

ei::GAType

Return

The reverse norm of ei.

Main.MinGal.revert — Method

revert(ei::GAType)::GAType

Function that computes the reverse of a multivector and return its result.

Arguments

ei::Multivector

Return

The reverse of ei.

Main.MinGal.right_contraction — Method

right_contraction(ei::GAType, ej::GAType)::GAType

Function that computes the right contraction of two multivectors and return its result.

Arguments

ei::GAType

ej::GAType

Return

The result GAType.

Main.MinGal.scalar — Method

scalar(bl::Blade)::Number||
scalar(mv::Multivector)::Number

Returns the scalar of the GAType

Arguments

bl::Blade

mv::Multivector

Return

The scalar

Main.MinGal.scalar_product — Method

scalar_product(ei, ej)::Number

Function that returns the Scalar Product between two GAType.

Arguments

ei::GAType

ej::GAType

Return

The result product.

Main.MinGal.scalar_product_basis — Method

scalar_product_basis(ei, ej)::Integer

Function that returns the Scalar Product between two basis blades.

Arguments

ei::Blade

ej::Blade

Return

The result Integer.

Main.MinGal.set_scalar — Method

set_scalar(mv, val, k)::Number ||
set_scalar(mv, val, ei)::Number

Function that changes the scalar value in index k from GAType mv. The index k follows the bit order basis.The ei value represents the blade that you want to change the scalar from.

Arguments

mv::GAType

val::Number : New Value

k::Integer : Index in bit order basis

||

mv::GAType

val::Number : New Value

ei::Blade : Index

Return

The scalar value.

Main.MinGal.undual — Method

undual(ei::GAType)::GAType

Function that computes the undual of a multivector and return its result. Just be careful with r > 0. It uses Polarity algorithm for invertible algebras and hodge dual for degenerate algebras.

Arguments

ei::Multivector

Return

The undual of ei.